Refactoring into separate ModuleScopeBuilder class
