+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/data/Type.java
@@ -39,6 +39,7 @@ import org.enso.interpreter.runtime.data.atom.AtomConstructor;
 import org.enso.interpreter.runtime.data.vector.ArrayLikeHelpers;
 import org.enso.interpreter.runtime.library.dispatch.TypesLibrary;
 import org.enso.interpreter.runtime.scope.ModuleScope;
+import org.enso.interpreter.runtime.scope.ModuleScopeBuilder;
 import org.enso.interpreter.runtime.util.CachingSupplier;
 import org.enso.pkg.QualifiedName;
 
@@ -47,7 +48,7 @@ import org.enso.pkg.QualifiedName;
 public final class Type extends EnsoObject {
 
   private final String name;
-  private @CompilerDirectives.CompilationFinal ModuleScope.Builder definitionScope;
+  private @CompilerDirectives.CompilationFinal ModuleScopeBuilder definitionScope;
   private final boolean builtin;
   private final Type supertype;
   private final Type eigentype;
@@ -59,7 +60,7 @@ public final class Type extends EnsoObject {
 
   private Type(
       String name,
-      ModuleScope.Builder definitionScope,
+      ModuleScopeBuilder definitionScope,
       Type supertype,
       Type eigentype,
       boolean builtin,
@@ -75,7 +76,7 @@ public final class Type extends EnsoObject {
 
   public static Type createSingleton(
       String name,
-      ModuleScope.Builder definitionScope,
+      ModuleScopeBuilder definitionScope,
       Type supertype,
       boolean builtin,
       boolean hasAllConstructorsPrivate) {
@@ -85,7 +86,7 @@ public final class Type extends EnsoObject {
   public static Type create(
       EnsoLanguage lang,
       String name,
-      ModuleScope.Builder definitionScope,
+      ModuleScopeBuilder definitionScope,
       Type supertype,
       Type any,
       boolean builtin,
@@ -131,7 +132,7 @@ public final class Type extends EnsoObject {
   }
 
   public void setShadowDefinitions(
-      EnsoLanguage lang, ModuleScope.Builder scope, boolean generateAccessorsInTarget) {
+      EnsoLanguage lang, ModuleScopeBuilder scope, boolean generateAccessorsInTarget) {
     if (builtin) {
       // Ensure that synthetic methods, such as getters for fields are in the scope.
       CompilerAsserts.neverPartOfCompilation();
